refactor: decompose git/operations.rs into focused modules
Split the 2,280-line operations.rs into 5 files by responsibility:
- naming.rs: path sanitization, branch names, project ID generation
- validation.rs: branch/arg validation, git directory checks
- status.rs: diff stats, worktree status, git stats aggregation
- health.rs: branch health metrics with pub(super) shared helpers
- overlaps.rs: file overlap detection across kilds

Updated mod.rs with re-exports so callers use cleaner paths
(e.g. crate::git::kild_branch_name instead of
crate::git::operations::kild_branch_name). Tests moved with
their code, no behavioral changes.
